requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
pydantic[email]
psycopg2-binary
alembic
jinja2
asyncpg
passlib
python-jose